Change Expect.equal failure message to show diffs
Instead of:

  "these strings really are not quite equal"
  ╷
  │ Expect.equal
  ╵
  "These strings aren't quite equal."

  { a = [1,3], b = "test?", c = Just 1 }
  ╷
  │ Expect.equal
  ╵
  { a = [1,2,3], b = "test", c = Just 3 }

It does:

  Unexpected differences:
  "[-t-]{+T+}hese strings[- really-] are[- -]n[-o-]{+'+}t quite equal{+.+}"

  Unexpected differences:
  { a = [1{+,2+},3], b = "test[-?-]", c = Just [-1-]{+3+} }

This isn't optimal, as colored diffs would be better.  But that is hard
because really the reporters should be in charge of formatting the
messages appropriately for their context.  That would require changing
both the node and html reporters, and the Expectation type to provide
the raw data for the failing values instead of a message String.